test: refactor binarizer and morphologyzer structure dir to lower complexity
