+++ tests/unit/image/components/transformer/binarizer/test_binarizer_binary.py
